Added "use_anti_afk" setting to enable/disable anti-AFK logic
This commit addresses an issue identified with bots occasionally breaking due to anti-AFK logic. To provide users with greater flexibility and control over their bots' behavior, a new setting called "use_anti_afk" has been added. By toggling this setting, users can now easily enable or disable the anti-AFK feature as needed.

To use this new setting, simply update your configuration file with the "use_anti_afk" parameter and set it to either "True" (to enable the anti-AFK logic) or "False" (to disable it). This enhancement aims to improve the stability and performance of bots by allowing users to decide whether or not to utilize the anti-AFK feature based on their specific requirements.