feat: damage calculation now correctly treats unique enchantment_ids as stackable spell effects. (#91)
I was playing around with damage calcs, and realized that it wasn't stacking enchanted blades of the same type with eachother. ie.
```
Stormblade
Stormblade (Sharpened)
Stormblade (Gear Sharpened)
```
Would all be considered duplicates.

After looking at the spell effects, I realized that the field `enchantment_spell_template_id ` is what we need to consider to allow these to stack properly. A SpellEffect is only a duplicate if both the template ID and the enchantment ID match. Conveniently, this is already in the cut-down attributes, in index location 4. Anyway, I attached a screensot of me testing this with the 3 stormblades above, and seeing that it calculates correctly.

![spell-effect-stack.jpg](/attachments/a62ac30e-4640-4c4e-928f-9be13c162b68)

![spell-effect-stack-deckjpg.jpg](/attachments/89cfbba9-6834-4ae7-a50e-f4c57c05908f)

maybe one of these days I'll try rounding in a bunch of places to see where the rounding is being done server side... So we can finally fix that

Co-authored-by: slackaduts <slackaduts@noreply.codeberg.org>
Reviewed-on: https://codeberg.org/Deimos-Wizard101/Deimos-Wizard101/pulls/91
Co-authored-by: Evan Hruskar <hruskar.evan@gmail.com>
Co-committed-by: Evan Hruskar <hruskar.evan@gmail.com>